mdcode: move the custom type declarations into their own file
`kc_actions.ts` held both the encoding of an action and the declaration
of the entry and aspect types that encoding writes to. Those are
different audiences. Someone deciding whether a type should be built-in
needs the id, the display name, and the metadata template; they do not
need the emitter or the reader, and they should not have to find the
declaration inside a file named after one construct.

`kc_custom_types.ts` is now the list of what kcmd creates rather than
references. It declares each type, provisions the whole list at init,
and names them; adding another custom type is one appended record, with
no other file in the directory changing. Its header states what a
Dataplex type owner has to do to make one built-in and what stays put
when they do.

`kc_actions.ts` keeps the part that is specific to actions: the model to
entries, and entries back to the model. It names its type through the
registry, so making the action type built-in no longer touches it.
