+++ toolbox/mdcode/src/libts/semantic/kc_actions.ts
@@ -1,32 +1,23 @@
-// How a model's ACTIONS are persisted in Knowledge Catalog.
+// How a model's ACTIONS are encoded in Knowledge Catalog.
 //
-// This file holds the whole encoding, and it exists to be replaced.
-//
-// Every other construct in the model maps to a BUILT-IN system type under
-// `projects/dataplex-types/locations/global`. A model, entity, or metric gets
-// its own entry (`semantic-model` / `semantic-entity` / `semantic-metric`); a
-// relationship becomes a `schema-join` entry link; the built-in `schema` and
-// `guidelines` aspects carry the rest. There is no built-in type for an action.
-// Until there is, this file defines a CUSTOM pair, an entry type and an aspect
-// type both named `semantic-action`, provisioned in the destination project at
-// `global` by `kcmd init --semantic-model`. An action is then published exactly
-// the way a metric is: one entry per action, parented to the model anchor,
-// carrying one aspect that holds the executor and the typed parameters.
+// An action is published exactly the way a metric is: one entry per action,
+// parented to the model anchor, carrying one aspect that holds the executor and
+// the typed parameters. What differs is the type. Every other construct has a
+// built-in system type, and an action does not, so it uses the custom
+// `semantic-action` pair DECLARED IN kc_custom_types.ts and created there by
+// `kcmd init --semantic-model`. That file is the list of what is custom; this
+// one is only the encoding that fills the aspect.
 //
 // The custom pair is what makes an action explicit in the catalog. A search can
 // tell an action apart from anything else by its entry type, and the aspect's
 // fields are typed and queryable rather than prose a reader has to interpret.
 //
-// WHEN A BUILT-IN ACTION TYPE SHIPS. Change actionTypeHome to return the
-// system-type home the caller already passes for every other construct, then
-// delete ACTION_ASPECT_TYPE_SPEC, actionEntryTypeSpec, and the
-// provisionActionTypes call in `kcmd init`. Nothing else moves: the entry
-// shape, the entry ids, the owned prefix, the reader, and all four call sites
-// stay as they are.
+// WHEN A BUILT-IN ACTION TYPE SHIPS, follow the instructions at the top of
+// kc_custom_types.ts. Nothing in this file changes: the readers below match a
+// type by its id suffix, so they do not care which project it lives in.
 //
 // The call sites are `knowledge_catalog.ts` (emit the entries),
-// `kc_converter.ts` (read them back), `pull_kc.ts` (hydrate the aspect), and
-// `src/tool/commands.ts` (provision the two types at init).
+// `kc_converter.ts` (read them back), and `pull_kc.ts` (hydrate the aspect).
 //
 // `instructions` rides the action's own aspect rather than the built-in
 // `guidelines` aspect that an entity or metric uses. A pull derives which
@@ -34,301 +25,31 @@
 // action entry, whose type is custom and therefore in the destination project,
 // would ask for a `guidelines` aspect type that only exists under
 // `dataplex-types`. Keeping the field on the action's own aspect keeps the
-// whole encoding inside the one type this file provisions.
+// whole encoding inside the one type kc_custom_types.ts provisions.
 //
 // The helpers at the bottom duplicate a few lines from the modules above on
-// purpose. This module imports only the IR and the catalog client, so it can be
-// swapped or deleted as a unit.
+// purpose. This module imports only the IR, the entry shape and the type
+// registry, so it can be swapped or deleted as a unit.
 
-import {AspectType, CatalogClient, Entry, EntryType} from '../gcp/dataplex';
+import {Entry} from '../gcp/dataplex';
 
 import {Action, ActionParameter, AiContext, DATA_TYPES, Executor, SemanticModel} from './ir';
-
-// The bare id shared by the custom entry type and the custom aspect type. They
-// are separate collections, so one name serves both, the way the built-in
-// `semantic-metric` entry type and aspect type share theirs.
-export const ACTION_TYPE_ID = 'semantic-action';
-
-// Custom types are provisioned at `global` so an entry group in any region can
-// reference them, matching where the built-in system types live.
-const ACTION_TYPE_LOCATION = 'global';
-
-// Where a destination's action types live. Today that is the destination
-// project itself, because the types are custom and `kcmd init` creates them
-// there. See the file header for what changes when a built-in type ships.
-export function actionTypeHome(dest: {project: string}):
-    {project: string; location: string} {
-  return {project: dest.project, location: ACTION_TYPE_LOCATION};
-}
-
-
-// ---------------------------------------------------------------------------
-// The custom types themselves.
-// ---------------------------------------------------------------------------
-
-// The aspect that carries an action's mechanics.
-//
-// The executor is FLATTENED rather than nested one record per kind: an aspect
-// holds exactly one executor, and a flat record shows the reader which one at a
-// glance instead of three sibling records of which two are empty. `parameters`
-// mirrors ActionParameter, including the derived `isEntityRef` so a consumer
-// that has not loaded the model can still tell an object reference from a
-// scalar.
-//
-// Dataplex rejects a backwards-incompatible template change. A new field must
-// be APPENDED with a fresh index; renumbering an existing one breaks the update
-// for every project that already provisioned this type.
-export const ACTION_ASPECT_TYPE_SPEC: Omit<AspectType, 'name'> = {
-  displayName: 'Semantic Action',
-  description:
-      'A write operation defined on a semantic model: how it is executed, ' +
-      'and the inputs it takes.',
-  metadataTemplate: {
-    name: ACTION_TYPE_ID,
-    type: 'record',
-    recordFields: [
-      {
-        index: 1,
-        name: 'executorKind',
-        type: 'string',
-        constraints: {required: true},
-        annotations: {
-          displayName: 'Executor Kind',
-          description:
-              'How the action is executed: `mcp`, `rest`, or `grpc`. The ' +
-              'fields for that kind, and only those, are set.',
-        },
-      },
-      {
-        index: 2,
-        name: 'mcpServer',
-        type: 'string',
-        annotations: {
-          displayName: 'MCP Server',
-          description: 'Resource name of the MCP server hosting the tool.',
-        },
-      },
-      {
-        index: 3,
-        name: 'mcpTool',
-        type: 'string',
-        annotations: {
-          displayName: 'MCP Tool',
-          description: 'Name of the tool to call on the MCP server.',
-        },
-      },
-      {
-        index: 4,
-        name: 'restEndpoint',
-        type: 'string',
-        annotations: {
-          displayName: 'REST Endpoint',
-          description: 'URL the request is sent to.',
-        },
-      },
-      {
-        index: 5,
-        name: 'restMethod',
-        type: 'string',
-        annotations: {
-          displayName: 'REST Method',
-          description: 'HTTP method, for example POST.',
-        },
-      },
-      {
-        index: 6,
-        name: 'grpcService',
-        type: 'string',
-        annotations: {
-          displayName: 'gRPC Service',
-          description: 'Fully qualified name of the gRPC service.',
-        },
-      },
-      {
-        index: 7,
-        name: 'grpcMethod',
-        type: 'string',
-        annotations: {
-          displayName: 'gRPC Method',
-          description: 'Method on the gRPC service.',
-        },
-      },
-      {
-        index: 8,
-        name: 'parameters',
-        type: 'array',
-        arrayItems: {
-          name: 'parameter',
-          type: 'record',
-          recordFields: [
-            {
-              index: 1,
-              name: 'name',
-              type: 'string',
-              constraints: {required: true},
-              annotations: {displayName: 'Name'},
-            },
-            {
-              index: 2,
-              name: 'type',
-              type: 'string',
-              constraints: {required: true},
-              annotations: {
-                displayName: 'Type',
-                description:
-                    'The authored type: an entity name, or a scalar datatype.',
-              },
-            },
-            {
-              index: 3,
-              name: 'isEntityRef',
-              type: 'bool',
-              annotations: {
-                displayName: 'Is Entity Reference',
-                description:
-                    'True when `type` names an entity, so the parameter ' +
-                    'refers to an object rather than carrying a value.',
-              },
-            },
-          ],
-        },
-        annotations: {
-          displayName: 'Parameters',
-          description:
-              'The inputs the action takes, each typed by the ontology.',
-        },
-      },
-      {
-        index: 9,
-        name: 'instructions',
-        type: 'string',
-        annotations: {
-          displayName: 'Instructions',
-          description:
-              'Guidance for AI consumers (the action\'s ai_context ' +
-              'instructions).',
-        },
-      },
-    ],
-  },
-};
-
-// The entry type an action's entry carries. It requires the aspect above, so an
-// action entry cannot exist without its mechanics.
-export function actionEntryTypeSpec(dest: {project: string}):
-    Omit<EntryType, 'name'> {
-  return {
-    displayName: 'Semantic Action',
-    description: 'A write operation defined on a semantic model.',
-    requiredAspects: [{type: actionAspectTypeName(dest)}],
-  };
-}
+import {ACTION_TYPE_ID, customAspectKey, customAspectTypeName, customEntryTypeName} from './kc_custom_types';
 
 // Full resource name of the action entry type for a destination.
 export function actionEntryTypeName(dest: {project: string}): string {
-  const home = actionTypeHome(dest);
-  return `projects/${home.project}/locations/${home.location}/entryTypes/${
-      ACTION_TYPE_ID}`;
+  return customEntryTypeName(ACTION_TYPE_ID, dest);
 }
 
 // Full resource name of the action aspect type for a destination.
 export function actionAspectTypeName(dest: {project: string}): string {
-  const home = actionTypeHome(dest);
-  return `projects/${home.project}/locations/${home.location}/aspectTypes/${
-      ACTION_TYPE_ID}`;
+  return customAspectTypeName(ACTION_TYPE_ID, dest);
 }
 
 // Aspect-map key: the `project.location.type` reference form the client keys an
-// entry's aspects by (see dataplex._nameToTypeRef).
+// entry's aspects by.
 export function actionAspectKey(dest: {project: string}): string {
-  const home = actionTypeHome(dest);
-  return `${home.project}.${home.location}.${ACTION_TYPE_ID}`;
-}
-
-// What provisioning produced. `error` is a failure the caller must not ignore.
-// `denied` is the narrower case of lacking permission to create a type, which
-// is reported and survived: actions are one optional construct, and a project
-// whose models declare none never needs these types.
-export interface ProvisionResult {
-  error?: string;
-  denied?: string;
-}
-
-/**
- * Creates the two custom action types in a destination.
- *
- * Called from `kcmd init --semantic-model`, beside the entry group it already
- * provisions, so a push still writes nothing but entries. The aspect type is
- * created and then, if it is already there, patched: a project provisioned by
- * an earlier version of kcmd holds an older template, and pushing a field that
- * template lacks fails with an opaque parsing error. Dataplex rejects a
- * backwards-incompatible change, so the patch only ever adds appended fields.
- */
-export async function provisionActionTypes(
-    cat: CatalogClient, dest: {project: string}): Promise<ProvisionResult> {
-  const home = actionTypeHome(dest);
-
-  // Creating a type needs permissions that publishing entries does not. A
-  // caller who lacks them can still init, push and pull every model that
-  // declares no action, so the refusal is reported rather than fatal.
-  const denial = (what: string, status: number, message?: string):
-      ProvisionResult|undefined => status === 403 ?
-      {
-        denied: `no permission to create ${what} in project '${
-            home.project}' (${message || status}); ` +
-            `models that declare actions cannot be pushed until it exists`
-      } :
-      undefined;
-
-  const aspectLabel = `aspect type '${ACTION_TYPE_ID}'`;
-  const aspect = await cat.createAspectType(
-      home.project, home.location, ACTION_TYPE_ID, ACTION_ASPECT_TYPE_SPEC);
-  if (aspect.status === 409) {
-    const upd = await cat.updateAspectType(
-        home.project, home.location, ACTION_TYPE_ID, ACTION_ASPECT_TYPE_SPEC,
-        ['description', 'display_name', 'metadata_template']);
-    const refused = denial(aspectLabel, upd.status, upd.message);
-    if (refused) return refused;
-    if (upd.status !== 200)
-      return {error: `updating ${aspectLabel}: ${upd.message || upd.status}`};
-    const failed =
-        await cat.awaitOperation(upd.result, `updating ${aspectLabel}`);
-    if (failed) return {error: failed};
-  } else if (aspect.status !== 200) {
-    const refused = denial(aspectLabel, aspect.status, aspect.message);
-    if (refused) return refused;
-    return {
-      error: `creating ${aspectLabel}: ${aspect.message || aspect.status}`
-    };
-  } else {
-    // The entry type below names this aspect type in its required aspects, and
-    // the server rejects that while the create is still in flight, so the wait
-    // is load-bearing rather than tidiness.
-    const failed =
-        await cat.awaitOperation(aspect.result, `creating ${aspectLabel}`);
-    if (failed) return {error: failed};
-  }
-
-  // The entry type requires the aspect type above, so it is created second. An
-  // existing one is left alone rather than patched: unlike the aspect type it
-  // carries no template that grows over time, and its required-aspect list is
-  // what an entry already published depends on.
-  const entryLabel = `entry type '${ACTION_TYPE_ID}'`;
-  const entryType = await cat.createEntryType(
-      home.project, home.location, ACTION_TYPE_ID, actionEntryTypeSpec(dest));
-  if (entryType.status === 409) return {};
-  const refused = denial(entryLabel, entryType.status, entryType.message);
-  if (refused) return refused;
-  if (entryType.status !== 200) {
-    return {
-      error: `creating ${entryLabel}: ${entryType.message || entryType.status}`
-    };
-  }
-  // A push may follow immediately, and an entry naming a type that is still
-  // being created is rejected the same way, so wait here too.
-  const failed =
-      await cat.awaitOperation(entryType.result, `creating ${entryLabel}`);
-  return failed ? {error: failed} : {};
+  return customAspectKey(ACTION_TYPE_ID, dest);
 }
